third_party/WebKit/LayoutTests/css3/calc/font-expected.txt

Tests that CSS3 calc() can be used with the font property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("calc-size"), null).lineHeight is "63px"
PASS getComputedStyle(document.getElementById("calc-size"), null).fontSize is "24px"
PASS getComputedStyle(document.getElementById("calc-height"), null).lineHeight is "63px"
PASS getComputedStyle(document.getElementById("calc-height"), null).fontSize is "24px"
PASS getComputedStyle(document.getElementById("calc-both"), null).lineHeight is "63px"
PASS getComputedStyle(document.getElementById("calc-both"), null).fontSize is "24px"
PASS getComputedStyle(document.getElementById("calc-percent-size"), null).lineHeight is "63px"
PASS getComputedStyle(document.getElementById("calc-percent-size"), null).fontSize is "24px"
PASS successfullyParsed is true

TEST COMPLETE
